% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seshat_api.R
\name{get_seshat_api}
\alias{get_seshat_api}
\title{Recuperer des donnees via l'API Seshat}
\usage{
get_seshat_api(endpoint = "core/polities")
}
\arguments{
\item{endpoint}{Le nom de la categorie (ex: "core/polities")}
}
\description{
Recuperer des donnees via l'API Seshat
}
